remove helpers.js
